requirements.txt
numpy==1.16.3
pandas==0.23.4
tweepy==3.8.0
Flask==1.0.2
gunicorn==19.6.0
google-cloud-datastore==1.7.3
Flask-Session==0.3.1
Flask-Restless==0.17.0
Flask-Paginate==0.5.0